fix: fixed all virtualization
